Index

Titus.FragmentMatch

Titus.Ion

Titus.MzFeature

Titus.Precursor

Titus.Precursor

Titus.Precursor

Titus.Precursor

Titus.Residue

Titus.Transition

Titus.Transition

Titus.getBIonModifier

Titus.getFragIons

Titus.getIonMZ

Titus.getIonMZ

Titus.getIonSeries

Titus.getNearest

Titus.getPrecursors

Titus.getYIonModifier

Titus.matchPeaks

Titus.matchPeaks!

Titus.setFragmentMatch!

Titus.FragmentMatch — Type

FragmentMatch

Type that represents a match between a fragment ion and a mass spectrum peak

Fields

transition::Transition – Represents a fragment ion

intensity::Float32 – Intensity of the matching peak

match_mz::Float32 – M/Z of the matching empirical peak. NOT the transition mass and may differ from getMZ(transition) by some error

count::UInt8 – Number of matches (may want to count the number of matches if the spectrum is researched at different mass offsets, such as in a cross correlation score)

peak_int::Int64 – Index of the matching peak in the mass spectrum.

Examples

`FragmentMatch(transition::Transition, intensity::Float32, mass::Float32, count::UInt8, peak_ind::Int64) – default internal constructor

FragmentMatch() – constructor for null/empty precursor

GetterMethods

getMZ(f::FragmentMatch) = getMZ(f.transition)

getLow(f::FragmentMatch) = getLow(f.transition)

getHigh(f::FragmentMatch) = getHigh(f.transition)

getPrecID(f::FragmentMatch) = getPrecID(f.transition)

getCharge(f::FragmentMatch) = getCharge(f.transition)

getIsotope(f::FragmentMatch) = getIsotope(f.transition)

getIonType(f::FragmentMatch) = getIonType(f.transition)

getInd(f::FragmentMatch) = getInd(f.transition)

Titus.Ion — Type

Ion

Abstract type that represents an ion

Types that inherit from Ion should implement the following.

getMZFeature(ion::Ion) = ion.mz

getMZ(ion::Ion) = getMZ(getMZFeature(ion))

getLow(ion::Ion) = getLow(getMZFeature(ion))

getHigh(ion::Ion) = getHigh(getMZFeature(ion))

getPrecID(ion::Ion) = ion.prec_id

getCharge(ion::Ion) = ion.charge

getIsotope(ion::Ion) = ion.isotope

Titus.MzFeature — Type

MzFeature

Type that represents an m/z ratio with a ppm tolerance

Fields

mono::Float32 – MZ with upper and lower bounds given a ppm tolerance

low::Float32 – Identifier of the precursor ion (parent ion of the fragment/transition)

high::Float32 – Type of transition. For example 'b' for b ion or 'y' for y ion

Examples

MzFeature(mono::Float32; ppm::Float32 = Float32(20)) – default internal constructor

MzFeature() = MzFeature(Float32(0.0)) – constructor for a default/placeholder MzFeature

GetterMethods

getMZ(mzfeature::MzFeature) = mzfeature.mono

getLow(mzfeature::MzFeature) = mzfeature.low

getHigh(mzfeature::MzFeature) = mzfeature.high

Titus.Precursor — Type

Precursor <: Ion

Type that represents a precursor (A peptide parent ion)

Fields

residues::Vector{Residue} – List of amino acid residues of the precursor in their appropriate order

mz::MzFeature – MZ with upper and lower bounds given a ppm tolerance

prec_id::UInt32 – Identifier of the precursor ion (parent ion of the fragment/transition)

charge::UInt8 – Charge of the fragment ion

isotope::UInt8 – Difference in number of neutrons from the monoisotopic fragment

pep_id::UInt32 – Identifier of the peptide from which precursor is derived

Examples

Precursor(residues::Vector{Residue}, mz::Float32, charge::UInt8, isotope::UInt8, pep_id::UInt32, prec_id::UInt32; ppm = Float32(20)) – default internal constructor

Precursor() – constructor for null/empty precursor

Precursor(residues::Vector{Residue}, charge::UInt8, isotope::UInt8 = UInt8(0), pep_id::UInt32 = UInt32(0), prec_id::UInt32 = UInt32(0) – Constructor that calculates mz without having to supply it

Precursor(sequence::String; mods_dict::Dict{String, Float32} = Dict{String, Float32}(), charge::UInt8 = UInt8(2), isotope::UInt8 = UInt8(0), pep_id::UInt32 = UInt32(0), prec_id::UInt32 = UInt32(0)) – Constructor that accepts a string representation of a peptide

GetterMethods

getResidues(precursor::Precursor) = precursor.residues

Titus.Precursor — Method

Precursor()
Constructor for an "empty" or "default" precursor

Titus.Precursor — Method

Precursor(sequence::String, mods_dict::Dict{String, Float32}, charge::Int32, isotope::Int32 = Int32(0), prec_id::Int32 = Int32(0), pep_id::Int32 = Int32(0))

Alternate constructor for the Precursor struct. Can accept a string representation of a peptide and a mods_dict and convert to residues Array{Residue, 1}. (link to Precusor)

Titus.Precursor — Method

Precursor(residues::Array{Residue, 1}, charge::Int32, isotope::Int32 = Int32(0), prec_id::Int32 = Int32(0), pep_id::Int32 = Int32(0))

Constructor for the Precursor struct. Given a list of amino acid residues, a charge, and an isotope state, makes a precursor object with the correct mz. (link to Precursor)

Titus.Residue — Type

Residue

Type that represents a (potentially modified) amino acid within a peptide

Fields

mass:Float32 – mass of the amino acid

Examples

Residue(aa::AA) – default constructor

Residue(aa::Char) = Residue(AA(aa))

Residue(aa::AA, mod::Mod) = Residue(getMass(mod)+getMass(aa))

Residue(residue::String, mods_dict::Dict{String, Float32}) = Residue(AA(residue[1]), Mod(residue, mods_dict))

Residue(residue::String) = Residue(residue, Dict{String, Float32}())

Residue(residue::String, mod_mass::Float32) = Residue(getMass(AA(residue[1])) + mod_mass)

Residue(residue::Char, mod_mass::Float32) = Residue(getMass(AA(residue)) + mod_mass)

Getter methods

getMass(residue::Residue) = residue.mass

See Also

getResidues(sequence::String, mods_dict::Dict{String, Float32} = default_mods) – Gets a vector of residues given a string representation of an amino acid sequence

Titus.Transition — Type

Transition <: Ion

Type that represents transition (fragment ion of a peptide)

Fields

mz::MzFeature – MZ with upper and lower bounds given a ppm tolerance

prec_id::UInt32 – Identifier of the precursor ion (parent ion of the fragment/transition)

ion_type::Char – Type of transition. For example 'b' for b ion or 'y' for y ion

ind::UInt8 – Position of fragment ion with reference to parent ion. (A b5+2 ion should have an ind equal to 5)

charge::UInt8 – Charge of the fragment ion

isotope::UInt8 – Difference in number of neutrons from the monoisotopic fragment

Examples

Transition(frag_mz::Float32, prec_id::UInt32, ion_type::Char, ind::UInt8, charge::UInt8, isotope::UInt8; ppm = Float32(20)) – default internal constructor

Transition(residues::Vector{Residue}; ion_type::Char = 'y', charge::UInt8 = UInt8(1), ind::UInt8 = UInt8(length(residues)), isotope::UInt8 = UInt8(0), prec_id::UInt32 = UInt32(0)) – constructor that calculates the appropriate mz

GetterMethods

getIonType(transition::Transition) = transition.ion_type

getInd(transition::Transition) = transition.ind

Titus.Transition — Method

Transition(residues::Vector{Residue}; ion_type::Char = 'y', charge::UInt8 = UInt8(1), ind::UInt8 = UInt8(length(residues)), isotope::UInt8 = UInt8(0), prec_id::UInt32 = UInt32(0))

Constructor for the `Transition` struct. Given a list of amino acid residues, ion type (b or y), charge state, and isotopic state makes a transition with the correct mz.

Titus.getBIonModifier — Method

getBIonModifier(charge::UInt8)

Mass modification that needs to be added to b-ions

Titus.getFragIons — Method

getFragIons(residues::Vector{Residue}; charge::UInt8 = UInt8(1), isotope::UInt8 = UInt8(0), y_start::Int = 3, b_start::Int = 3)

Uses calls to getIonSeries to concatenate both the b and y ion series together in a single Vector{Float32}

Input

residues::Vector{Residue}: – List of amino acid residues in the peptide ion

charge::UInt8 – Charge of the fragment ions

b_start::Int=3 – Index of first ion the the b-ion series to compute.

y_start::Int=3 – Index of first ion the the y-ion series to compute.

isotope::UInt8=UInt8(0) – Diference in the number of isotopes from the monoisotopic ion.

Output

A Vector{Float32} wich each m/z in the ion series

Notes

The modifier argument ought to depend on the kind of ion. For a 'b' ion series PROTON*charge is appropriate,

but for a 'y' ion series, PROTON*charge + H2O would be appropriate.

Will not allow the index of the ion to be equal to or less than the charge. For example,

b2+2 ions could only be calculated in error and are therefore excluded even if start is set to 2.

If start exceeds length(residues)-1, then only the N-1 ion is calculated, that is,

the highest mass ion in the series.

Examples

#Gets the b3+1-b6+1 and y3+1-y6+1 ions

julia> getFragIons(reverse(getResidues("PEPTIDE")), b_start = 3, y_start = 3)
8-element Vector{Float32}:
 358.16083
 459.2085
 556.2612
 685.30383
 342.16595
 443.21365
 556.29767
 671.3246

See Also

Alternate convience methods

getFragIons(residues::Vector{Residue}; charge::UInt8 = UInt8(1), isotope::UInt8 = UInt8(0), y_start::Int = 3, b_start::Int = 3) - Default method

getFragIons(precursor::Precursor; charge::UInt8 = UInt8(1), isotope::UInt8 = UInt8(0), y_start::Int = 3, b_start::Int = 3) - Can supply a Precursor rather than a Vector{Residues} input

getFragIons(precursor::Precursor,charges::Vector{UInt8}, isotopes::Vector{UInt8}; y_start::Int = 3, b_start::Int = 3) - Gets b and y ion seriers for multiple charge and isotopic states

Titus.getIonMZ — Method

getIonMZ(residues::Vector{Residue}, ion_type::Char, charge::UInt8; isotope::UInt8 = UInt8(0))

Alternate getIonMZ method that chooses the correct mass modifier for 'b', 'y', and 'p' ions respectively.

Input

- `residues::Vector{Residue}`: -- List of amino acid residues in the peptide ion
- `ion_type::Char` -- Type of fragment ion. Currently supports, 'b', 'y', and 'p'. 
- `charge::UInt8` -- Charge of the ion
- `isotope::UInt8=UInt8(0)` -- Diference in the number of isotopes from the monoisotopic ion.

Notes

See main method
    getIonMZ(residues::Vector{Residue}, charge::UInt8; modifier::Float32 = PROTON*charge + H2O, isotope::UInt8 = UInt8(0))
for more details

Titus.getIonMZ — Method

getIonMZ(residues::Vector{Residue}, charge::UInt8; modifier::Float32 = PROTON*charge + H2O, isotope::UInt8 = UInt8(0))::Float32

Get the mz ratio of an ion

Input

residues::Vector{Residue}: – List of amino acid residues in the peptide ion

charge::UInt8 – Charge of the ion

modifier::Float32=PROTON*charge + H2O – Added to the mass of the ion

isotope::UInt8=UInt8(0) – Diference in the number of isotopes from the monoisotopic ion.

Output

A Float32 representing the mass-to-charge ratio (m/z) of an ion

Notes

The modifier argument ought to depend on the kind of ion. For B ions PROTONcharge is appropriate, but for 'y' or precursor ions, PROTONcharge + H2O would be appropriate.

Algorithm

Sum the amino acid residue masses, add modifier + isotope*NEUTRON and then divide the total by the charge.

Examples

#Gets the b6+1 ion MZ

julia> getIonMZ(getResidues("PEPTIDE")[1:6], UInt8(1), modifier = PROTON)
653.314f0

#Gets the y6+1 ion MZ

julia> getIonMZ(reverse(getResidues("PEPTIDE"))[1:6], UInt8(1))
703.3144f0

Titus.getIonSeries — Method

getIonSeries(residues::Vector{Residue}, charge::UInt8; start::Int = 3, modifier::Float32 = PROTON*charge + H2O, isotope::UInt8 = UInt8(0))

Gets the m/z's for an ion series as a Vector{Float32}.

Input

residues::Vector{Residue}: – List of amino acid residues in the peptide ion

charge::UInt8 – Charge of the fragment ions

start::Int=3 – Index of first ion the the series to compute.

isotope::UInt8=UInt8(0) – Diference in the number of isotopes from the monoisotopic ion.

Output

A Vector{Float32} wich each m/z in the ion series

Notes

The modifier argument ought to depend on the kind of ion. For a 'b' ion series PROTON*charge is appropriate,

but for a 'y' ion series, PROTON*charge + H2O would be appropriate.

Will not allow the index of the ion to be equal to or less than the charge. For example,

b2+2 ions could only be calculated in error and are therefore excluded even if start is set to 2.

If start exceeds length(residues)-1, then only the N-1 ion is calculated, that is,

the highest mass ion in the series.

Examples

#Gets the y3+2 through y6+2 ions

julia> getIonSeries(reverse(getResidues("PEPTIDE")), UInt8(2), start = 3)
4-element view(::Vector{Float32}, 3:6) with eltype Float32:
 188.58934
 239.11317
 287.63956
 352.16086

#Gets the b4+2 through b6+2 ions

julia> getIonSeries(getResidues("PEPTIDE"), UInt8(2), start = 4, modifier = PROTON*UInt8(2))
3-element view(::Vector{Float32}, 4:6) with eltype Float32:
 213.10518
 269.6472
 327.16064

Titus.getNearest — Method

getNearest(transition::Transition, masses::Vector{Union{Missing, Float32}}, peak::Int; δ = 0.01)

Finds the peak (index of masses) nearest in mass to the transition but still within the tolerance (getLow(transition)<masses[peak]<getHigh(transition)). Starts searching at initially supplied peak and increases the index until outside the tolerance. There could be multiple peaks within the tolerance, and this function selects the one with the lowest mass error to the fragment ion.

Input

transition::Transition: – Represents a fragment ion

masses::Vector{Union{Missing, Float32}} – Mass list from a centroided mass spectrum. MUST BE SORTED IN ASCENDING ORDER.

peak::Int – An index for a mass in masses

δ – A mass offset that can be applied to each mass in masses

Output

An Int representing the index of the m/z in masses nearest in m/z to that of the fragment m/z.

Notes

It is assumed that masses is sorted in ascending order.

In practice, when called from matchPeaks!, masses[peak] will already be within the fragment m/z tolerance.

Usually there will not be another peak in masses where this is true, but it is possible for multiple peaks to fall within the tolerance. The purpose of this function is to select the best peak (closes in mass) when this happens.

Algorithm

Examples

Titus.getPrecursors — Method

getPrecursors(residues::Vector{Residue}; charges::Vector{UInt8} = UInt8[1,2],isotopes::Vector{UInt8}=UInt8[0],pep_id::UInt32=UInt32(0))

Alternate constructor for the Precursor struct that Can accept a string representation of a peptide

Input

Output

Notes

(link to getResidues())

Titus.getYIonModifier — Method

getYIonModifier(charge::UInt8)

Mass modification that needs to be added to y-ions

Titus.matchPeaks! — Method

function matchPeaks!(matches::Vector{FragmentMatch}, Transitions::Vector{Transition}, masses::Vector{Union{Missing, Float32}}, intensities::Vector{Union{Missing, Float32}}, δ::Float64)

Finds the best matching peak in a mass spectrum for each transition/fragment ion supplied if the match is within the fragment tolerance. Adds each FragmentMatch to matches if not already present. Otherwise, modifies an existing match (see setFragmentMatch!).

Input

matches::Vector{FragmentMatch}: – A list representing fragment ions that match peaks in the mass spectrum (masses)

Transitions::Vector{Transition – A list of fragment ions to search for in the spectrum (masses). MUST BE SORTED IN ASCENDING ORDER BY getMZ(transition)

masses::Vector{Union{Missing, Float32}} – Mass list from a centroided mass spectrum. MUST BE SORTED IN ASCENDING ORDER.

intensities::Vector{Union{Missing, Float32}} – The intensity list from a centroided mass spectrum. Must be the same length as masses

δ::Float64 – A mass offset that can be applied to each mass in masses

Output

Modifies matches[match] if match is <= lenth(matches). Otherwise adds a new FragmentMatch at matches[match]

Notes

Updating a match in matches could be useful if researching the same spectra many times at different mass offsets with matchPeaks! This could be done to calculate a cross correlatoin score for example. If a spectrum is only searched once, then matches should only be added to matches and existing ones never modified.

The fragment tolerance is specified by each Transition. A Transition<:Ion has a field mz::MzFeature which specifies the monoisotopic mass and also upper and lower bounds (the tolerance). See getLow(ion::Ion) and getHigh(ion::Ion). This is why the user need not supply a fragment tolerance to matchPeaks!

masses and intensities contain type unions Union{Missing, Float32}. This method does nothing to check for Missing values, and indeed, it is assumed that there are none, and the presence of any Missing values will cause an error. The reason for the type union is an idiosyncracy of the Arrow.jl package and how it implements nested data types in Arrow files.

Algorithm

Given a list of fragment ions and a centroided mass spectrum both sorted by m/z, it is efficient to search the spetrum for matches in a "single pass" through the spectrum. If there are T transitions and P peaks should be O(T+P). If there are multiple peaks within the tolerance for a given fragment ion, the peak closest in m/z to the fragment ion is chosen. It is possible to assign the same peak to multiple fragment ions, but each fragment ion is only assigned to 0 or 1 peaks.

Examples

Titus.matchPeaks — Method

function matchPeaks!(matches::Vector{FragmentMatch}, Transitions::Vector{Transition}, masses::Vector{Union{Missing, Float32}}, intensities::Vector{Union{Missing, Float32}}, δ::Float64)

A wrapper for calling matchPeaks at different to search spectra at a list of mass offset. Each all to matchPeaks Finds the best matching peak in a mass spectrum for each transition/fragment ion supplied if the match is within the fragment tolerance. Adds each FragmentMatch to matches if not already present. Otherwise, modifies an existing match (see setFragmentMatch!). (see matchPeaks for additional details)

Input

matches::Vector{FragmentMatch}: – A list representing fragment ions that match peaks in the mass spectrum (masses)

Transitions::Vector{Transition – A list of fragment ions to search for in the spectrum (masses). MUST BE SORTED IN ASCENDING ORDER BY getMZ(transition)

masses::Vector{Union{Missing, Float32}} – Mass list from a centroided mass spectrum. MUST BE SORTED IN ASCENDING ORDER.

intensities::Vector{Union{Missing, Float32}} – The intensity list from a centroided mass spectrum. Must be the same length as masses

δ::Float64 – A mass offset that can be applied to each mass in masses

Output

Modifies matches[match] if match is <= lenth(matches). Otherwise adds a new FragmentMatch at matches[match]

Notes

Searching a mass spectrum many times at different mass offsets could be useful for caculating cross correlation scores.

Algorithm

See `matchPeaks`

Examples

Titus.setFragmentMatch! — Method

setFragmentMatch!(matches::Vector{FragmentMatch}, match::Int, transition::Transition, mass::Float32, intensity::Float32, peak_ind::Int64)

Adds a FragmentMatch to matches if match is not an index in matches, otherwise, updates the match.

Input

hits::Vector{FragmentMatch}: – Represents a fragment ion

match::Int – Index of the match. Must be <=N+1 where N is length(hits)

mass::Float32 – m/z of the emperical peak matched to the transition

intensity::Float32 – intensity of the emperical peak matched to the transition

peak_ind – unique index of the emperical peak matched to the transition

Output

Modifies matches[match] if match is <= lenth(matches). Otherwise adds a new FragmentMatch at matches[match]

Notes

Updating a match in matches could be useful if researching the same spectra many times at different mass offsets with matchPeaks! This could be done to calculate a cross correlatoin score for example. If a spectrum is only searched once, then matches should only be added to matches and existing ones never modified.

Recording the peak_ind could be useful, for example, "chimeric" scoring of a spectrum from a Vector{FragmentMatch} type. The best scoring precursor would have fragments matching to known peak_ind. The Vector{FragmentMatch} could be rescored but excluding FragmentMatches corresponding to those peak_ind's. This would enable a simple chimeric spectra scoring that would not involve completely researching the spectrum (making an additional call to matchPeaks).

Algorithm

Examples